+++ packages/backend/src/nest/storage/storage.service.spec.ts
@@ -206,6 +206,11 @@ describe('StorageService', () => {
         value: orgProcessPlatform,
       })
     })
+
+    it('db address should be the same on all platforms', () => {
+      const dbAddress = StorageService.dbAddress({ root: 'zdpuABCDefgh123', path: 'channels.general_abcd' })
+      expect(dbAddress).toEqual(`/orbitdb/zdpuABCDefgh123/channels.general_abcd`)
+    })
   })
 
   describe('Channels', () => {
@@ -237,6 +242,29 @@ describe('StorageService', () => {
       expect(eventSpy).toBeCalledWith('channelDeletionResponse', {
         channelId: channelio.id,
       })
+    })
+
+    it('subscribes to pubsub on channel creation', async () => {
+      // @ts-expect-error 'subscribeToPubSub' is private
+      const subscribeToPubSubSpy = jest.spyOn(storageService, 'subscribeToPubSub')
+      await storageService.init(peerId)
+      await storageService.subscribeToChannel(channelio)
+      const db = storageService.publicChannelsRepos.get(channelio.id)?.db
+      expect(db).not.toBe(undefined)
+      if (!db) return // TS complaining
+      const channelsDbAddress = storageService.channels?.address
+      // @ts-expect-error 'certificates' is private
+      const certificatesDbAddress = storageService.certificates.address
+      expect(channelsDbAddress).not.toBeFalsy()
+      expect(certificatesDbAddress).not.toBeFalsy()
+      expect(subscribeToPubSubSpy).toBeCalledTimes(2)
+      // Storage initialization:
+      expect(subscribeToPubSubSpy).toHaveBeenNthCalledWith(1, [
+        StorageService.dbAddress(channelsDbAddress),
+        StorageService.dbAddress(certificatesDbAddress),
+      ])
+      // Creating channel:
+      expect(subscribeToPubSubSpy).toHaveBeenNthCalledWith(2, [StorageService.dbAddress(db.address)])
     })
   })