+++ packages/desktop/src/renderer/components/ContextMenu/ContextMenu.stories.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import { ComponentStory, ComponentMeta } from '@storybook/react'
-
 import { withTheme } from '../../storybook/decorators'
 
 import { ContextMenu } from './ContextMenu.component'