+++ packages/state-manager/src/sagas/identity/identity.master.saga.ts
@@ -5,6 +5,7 @@ import { registerCertificateSaga } from './registerCertificate/registerCertifica
 import { saveOwnerCertToDbSaga } from './saveOwnerCertToDb/saveOwnerCertToDb.saga'
 import { registerUsernameSaga } from './registerUsername/registerUsername.saga'
 import { savedOwnerCertificateSaga } from './savedOwnerCertificate/savedOwnerCertificate.saga'
+import { verifyJoinTimestampSaga } from './verifyJoinTimestamp/verifyJoinTimestamp.saga'
 
 export function* identityMasterSaga(socket: Socket): Generator {
   yield all([
@@ -12,5 +13,6 @@ export function* identityMasterSaga(socket: Socket): Generator {
     takeEvery(identityActions.registerCertificate.type, registerCertificateSaga, socket),
     takeEvery(identityActions.saveOwnerCertToDb.type, saveOwnerCertToDbSaga, socket),
     takeEvery(identityActions.savedOwnerCertificate.type, savedOwnerCertificateSaga, socket),
+    takeEvery(identityActions.verifyJoinTimestamp.type, verifyJoinTimestampSaga),
   ])
 }